LOKI_* for Makefile vars; remote stagenet
Loki technically has a stagenet supported in the code, but it isn't an
active network, so don't display it as an option in the wallet.

Rename the LOKI_VERSION and LOKI_ALPHA/BETA variables (from MONERO_*) in
the Makefile.  (Leave most of the other MONERO_ defined and monero_
functions as is since they are mostly unchanged from Loki).