Fix wallet address handling
Loki testnet addresses are 2 chars longer than mainnet addresses, but
the code assumed 95 everywhere.

This fixes it to properly handle the address size by returning it from
the wallet address conversion function.

Also rename the poorly named monero_base58_public_key to
loki_wallet_address because it *isn't* a base58 public key: it's more
than that (it also has networkid and checksum in it).  It is, in fact, a
wallet address, so name it appropriately.