Remove multi account feature, unused ux_menu
This is unnecessary and undesirable as the LOKI wallet already supports
multiple accounts.  There is an open bug on the Monero account for
removing it, but Monero can't easily because it might be in use.  We
have no such issue.

Also remove unused ux_menu variable; together this trims 30 bytes off
the volatile state data.