Staking support
Adds support for recognizing the tx type when we `open_tx()` which then
lets us distinguish between a regular TX, stake, unlock, or LNS when
displaying the "Processing [...]" message.

When in staking mode, we allow the private key to be retrieved
unencrypted (since it is going to be public knowledge in the tx
already).  In other tx modes this isn't allowed.

Staking mode also requires that the recipient must equal the wallet
primary address, and thus doesn't have to display the address for
confirmation at all (just the stake amount).