Add debug-only network reset instruction
This lets the device be forcibly reset to a specific network type by the
companion software wallet.  This command is *only* permitted when the
wallet is compiled in debug mode.

This also allows putting the wallet into fakechain mode, which is needed
for network tests, and allows doing so without needing (or exposing) a
fakechain setting.